// src/models/ErrorResponseModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.12.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Standard error response model.
 * 
 * Attributes:
 *     status_code (int): HTTP status code of the error.
 *     detail (str): A detailed message describing the error.
 * @export
 * @interface ErrorResponseModel
 */
export interface ErrorResponseModel {
    /**
     * 
     * @type {number}
     * @memberof ErrorResponseModel
     */
    statusCode: number;
    /**
     * 
     * @type {string}
     * @memberof ErrorResponseModel
     */
    detail: string;
}

/**
 * Check if a given object implements the ErrorResponseModel interface.
 */
export function instanceOfErrorResponseModel(value: object): value is ErrorResponseModel {
    if (!('statusCode' in value) || value['statusCode'] === undefined) return false;
    if (!('detail' in value) || value['detail'] === undefined) return false;
    return true;
}

export function ErrorResponseModelFromJSON(json: any): ErrorResponseModel {
    return ErrorResponseModelFromJSONTyped(json, false);
}

export function ErrorResponseModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): ErrorResponseModel {
    if (json == null) {
        return json;
    }
    return {
        
        'statusCode': json['status_code'],
        'detail': json['detail'],
    };
}

export function ErrorResponseModelToJSON(json: any): ErrorResponseModel {
    return ErrorResponseModelToJSONTyped(json, false);
}

export function ErrorResponseModelToJSONTyped(value?: ErrorResponseModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status_code': value['statusCode'],
        'detail': value['detail'],
    };
}